+++ tests/test_mprm_websocket.py
@@ -1,8 +1,6 @@
 import time
 
 import pytest
-from requests.exceptions import ConnectionError
-from urllib3.connection import ConnectTimeoutError
 from websocket import WebSocketApp
 
 from devolo_home_control_api.backend.mprm_rest import MprmRest
@@ -71,18 +69,18 @@ class TestMprmWebsocket:
         assert spy.call_count == 1
 
     @pytest.mark.usefixtures("mock_mprmwebsocket_websocketapp")
-    def test__try_reconnect(self, mocker):
+    def test__try_reconnect(self, request, mocker):
         spy = mocker.spy(time, "sleep")
-        self.mprm._ws = ConnectTimeoutError
+        self.mprm._test = request.node.name
         self.mprm._local_ip = self.gateway["local_ip"]
         self.mprm._try_reconnect(0.1)
         spy.assert_called_once_with(0.1)
 
     @pytest.mark.usefixtures("mock_mprmwebsocket_websocketapp")
-    def test__try_reconnect_with_detect(self, mocker):
+    def test__try_reconnect_with_detect(self, request, mocker):
         spy_sleep = mocker.spy(time, "sleep")
         spy_detect_gateway = mocker.spy(StubMprmWebsocket, "detect_gateway_in_lan")
-        self.mprm._ws = ConnectionError
+        self.mprm._test = request.node.name
         self.mprm._local_ip = self.gateway["local_ip"]
         self.mprm._try_reconnect(4)
         spy_sleep.assert_called_once_with(1)